refactor(NumberUtil.java) : 숫자 유틸 패키지 이동
