+++ src/main/java/lotto/domain/NumberUtil.java
@@ -1,4 +1,4 @@
-package utils;
+package lotto.domain;
 
 import java.util.ArrayList;
 import java.util.List;
