+++ src/test/java/lotto/domain/NumberUtilTest.java
@@ -1,4 +1,4 @@
-package utils;
+package lotto.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
